Use correct hascode method of Byte

+++ bvm/ballerina-core/src/main/java/org/ballerinalang/model/values/BByte.java
@@ -95,6 +95,6 @@ public final class BByte extends BValueType implements BRefType<Byte> {
 
     @Override
     public int hashCode() {
-        return Integer.hashCode(value);
+        return Byte.hashCode(value);
     }
 }
